fix(askar): throw error if imported wallet exists

+++ packages/askar/src/wallet/AskarWallet.ts
@@ -12,6 +12,7 @@ import {
   FileSystem,
   WalletNotFoundError,
   KeyDerivationMethod,
+  WalletImportPathExistsError,
 } from '@aries-framework/core'
 // eslint-disable-next-line import/order
 import { Store } from '@hyperledger/aries-askar-shared'
@@ -111,6 +112,11 @@ export class AskarWallet extends AskarBaseWallet {
       })
     }
     try {
+      // Make sure path exists before creating the wallet
+      if (filePath) {
+        await this.fileSystem.createDirectory(filePath)
+      }
+
       this._store = await Store.provision({
         recreate: false,
         uri: askarWalletConfig.uri,
@@ -280,7 +286,12 @@ export class AskarWallet extends AskarBaseWallet {
     }
 
     try {
-      // This method ensures that destination directory is created
+      // Export path already exists
+      if (await this.fileSystem.exists(destinationPath)) {
+        throw new WalletExportPathExistsError(
+          `Unable to create export, wallet export at path '${exportConfig.path}' already exists`
+        )
+      }
       const exportedWalletConfig = await this.getAskarWalletConfig({
         ...this.walletConfig,
         storage: { type: 'sqlite', path: destinationPath },
@@ -289,12 +300,8 @@ export class AskarWallet extends AskarBaseWallet {
       // Close this wallet before copying
       await this.close()
 
-      // Export path already exists
-      if (await this.fileSystem.exists(destinationPath)) {
-        throw new WalletExportPathExistsError(
-          `Unable to create export, wallet export at path '${exportConfig.path}' already exists`
-        )
-      }
+      // Make sure destination path exists
+      await this.fileSystem.createDirectory(destinationPath)
 
       // Copy wallet to the destination path
       await this.fileSystem.copyFile(sourcePath, destinationPath)
@@ -311,13 +318,13 @@ export class AskarWallet extends AskarBaseWallet {
 
       await this._open(this.walletConfig)
     } catch (error) {
-      if (error instanceof WalletExportPathExistsError) throw error
-
       const errorMessage = `Error exporting wallet '${this.walletConfig.id}': ${error.message}`
       this.logger.error(errorMessage, {
         error,
         errorMessage: error.message,
       })
+
+      if (error instanceof WalletExportPathExistsError) throw error
 
       throw new WalletError(errorMessage, { cause: error })
     }
@@ -332,8 +339,15 @@ export class AskarWallet extends AskarBaseWallet {
     }
 
     try {
-      // This method ensures that destination directory is created
       const importWalletConfig = await this.getAskarWalletConfig(walletConfig)
+
+      // Import path already exists
+      if (await this.fileSystem.exists(destinationPath)) {
+        throw new WalletExportPathExistsError(`Unable to import wallet. Path '${importConfig.path}' already exists`)
+      }
+
+      // Make sure destination path exists
+      await this.fileSystem.createDirectory(destinationPath)
 
       // Copy wallet to the destination path
       await this.fileSystem.copyFile(sourcePath, destinationPath)
@@ -354,6 +368,13 @@ export class AskarWallet extends AskarBaseWallet {
         error,
         errorMessage: error.message,
       })
+
+      if (error instanceof WalletImportPathExistsError) throw error
+
+      // Cleanup any wallet file we could have created
+      if (await this.fileSystem.exists(destinationPath)) {
+        await this.fileSystem.delete(destinationPath)
+      }
 
       throw new WalletError(errorMessage, { cause: error })
     }
@@ -387,13 +408,9 @@ export class AskarWallet extends AskarBaseWallet {
   private async getAskarWalletConfig(walletConfig: WalletConfig) {
     const { uri, path } = uriFromWalletConfig(walletConfig, this.fileSystem.dataPath)
 
-    // Make sure path exists before creating the wallet
-    if (path) {
-      await this.fileSystem.createDirectory(path)
-    }
-
     return {
       uri,
+      path,
       profile: walletConfig.id,
       // FIXME: Default derivation method should be set somewhere in either agent config or some constants
       keyMethod: keyDerivationMethodToStoreKeyMethod(
